Fix compiler warnings

+++ vm/src/load.c
@@ -6,7 +6,7 @@
 /*   By: dlancar <dlancar@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/12/12 14:36:38 by dlancar           #+#    #+#             */
-/*   Updated: 2017/01/25 15:38:34 by dlancar          ###   ########.fr       */
+/*   Updated: 2017/01/26 11:45:34 by dlancar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,7 +73,7 @@ t_program		*load_program(const char *path, int id)
 	off_t		size;
 
 	size = get_file_size(path);
-	if (size < sizeof(t_header))
+	if (size < (off_t)sizeof(t_header))
 		ft_error_msg("File %s is too small to be a champion\n", path);
 	content = get_file_content(path);
 	if (!(prog = ft_memalloc(sizeof(t_program))))
